1.74.7 (5.3.8p2)
Build-Id: 90583148-ea0b-46e0-8d67-d537430cbb72
Build-Sha256: 46f9f4bcafe09b15c9baa0639b6751dcaa17c790a005ca3160c9631fa16de6fe

+++ Assembly-CSharp/GameSessionData.cs
@@ -80,7 +80,7 @@ public class GameSessionData
 		serverIP = (string)gameSessionData["serverIP"];
 		profileID = (int)gameSessionData["profileID"];
 		planetID = (int)gameSessionData["planetID"];
-		gameMode = (MVGameMode)gameSessionData["gameMode"];
+		gameMode = (MVGameMode)(int)gameSessionData["gameMode"];
 		language = (string)gameSessionData["language"];
 		embedded = (bool)gameSessionData["embedded"];
 		token = (string)gameSessionData["token"];
